uv/tests: add new 'ecosystem' integration test
This adds a new top-level directory with bare-bones directories for a
sampling of ecosystem projects. The idea is for each directory to have
enough that `uv lock` can run.

The point of these tests is to 1) ensure resolution works in common
cases and 2) track changes to resolutions (and the lock file) in real
world projects.

Unfortunately, it does look like in some cases, re-running `uv lock`
results in changes to the lock file. For those cases, I've disabled the
deterministic checking in exchange for getting the lock files tracked
in tests. I haven't investigated yet whether either of #5733 or #5887
fix the deterministic problem.

There is probably a better way to go about integrating ecosystem
projects. In particular, it would be really nice if there was a good
flow for upgrading ecosystem packages to their latest version. The main
complexity is that some projects require edits to their `pyproject.toml`
(or a complete migration from non-`pyproject.toml` to `pyproject.toml`).
Although, the projects added here in this initial set were limited to
those that didn't require any changes.